🛠 DB (about Suggestion)

+++ src/main/java/com/github/pdgs/MyPengAPI/suggestion/entity/Suggestion.java
@@ -19,7 +19,7 @@ public class Suggestion {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int suggestionId;
+    private Long suggestionId;
 
     @Column(nullable = false, unique = true)
     private Long writerId;
